Reworked css structuring
styles.css was originally made for the mockup, and everything else was kinda piggybacking off of it. I wanted to rework the css structure, but I'm also trying to preserve the original mockup as much as I can since it represents a prototype of my software dev project. As such I'm reorganizing such that styles.css is only used by the mockup to keep that project isolated from everything else.